requirements.txt
Django==4.2.5
asgiref==3.7.2
djangorestframework==3.14.0
pip==23.2.1
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.2